C# Algorithms
This C# programme reads input from the user and stores their name in the
Name variable to use it later. It uses the Time API to get the current
time and stores that in the Time variable and uses that later aswell. It
then writes the two variables to the screen and then uses an algorithm
to pause the programme until AnyKey is pressed to continue.